requirements.txt
asgiref==3.4.1
Django==3.2.20
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.1.1